feat(be): 업체 리포지토리의 ID 타입을 Long으로 수정 #7

+++ backend/dagagomap/src/main/java/kr/dagagomap/repository/CompanyJpaRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import kr.dagagomap.entity.Company;
 
-public interface CompanyJpaRepository extends JpaRepository<Company, Integer> {
+public interface CompanyJpaRepository extends JpaRepository<Company, Long> {
 
 }
